test : 폼 변환 테스트 추가
