+++ src/test/java/stringcalculator/calculator/PostfixCalculatorTest.java
@@ -174,66 +174,4 @@ class PostfixCalculatorTest {
         );
     }
 
-    @ParameterizedTest(name = "[{index}] 잘못된 입력 기호 = {0}")
-    @MethodSource("getSignsToTestWhenWrongOperatorThenThrowException")
-    @DisplayName("올바르지 않은 기호가 포함된 후위 계산식 실패 예외처리 테스트")
-    void whenWrongOperatorThenThrowException(String operator) {
-        assertThatExceptionOfType(RuntimeException.class)
-                .as("Test operator =  %s", operator)
-                .isThrownBy(() -> calculator.getResult(splitWithoutSpace(operator)))
-                .withMessage(FORMULA_NULL_ERROR.getCode() + " :: " + FORMULA_NULL_ERROR.getDesc());
-    }
-
-    static Stream<Arguments> getSignsToTestWhenWrongOperatorThenThrowException() {
-        return Stream.of(
-                Arguments.arguments("!"),
-                Arguments.arguments("@"),
-                Arguments.arguments("#"),
-                Arguments.arguments("$"),
-                Arguments.arguments("%"),
-                Arguments.arguments("^"),
-                Arguments.arguments("&"),
-                Arguments.arguments("_"),
-                Arguments.arguments("="),
-                Arguments.arguments("\\"),
-                Arguments.arguments("|"),
-                Arguments.arguments("||"),
-                Arguments.arguments("~"),
-                Arguments.arguments("`"),
-                Arguments.arguments("'"),
-                Arguments.arguments("\""),
-                Arguments.arguments("["),
-                Arguments.arguments("]"),
-                Arguments.arguments("{"),
-                Arguments.arguments("}"),
-                Arguments.arguments(":"),
-                Arguments.arguments(";"),
-                Arguments.arguments(","),
-                Arguments.arguments("."),
-                Arguments.arguments("<"),
-                Arguments.arguments(">"),
-                Arguments.arguments("?")
-        );
-    }
-
-    @ParameterizedTest(name = "[{index}] 잘못된 입력 기호 = {0}")
-    @MethodSource("getAlphabetsToTestWhenWrongOperatorThenThrowException")
-    @DisplayName("올바르지 않은 알파벳이 포함된 후위 계산식 실패 예외처리 테스트")
-    void whenAlphaIsOperatorThenThrowException(String operator) {
-        assertThatExceptionOfType(RuntimeException.class)
-                .as("Test operator =  %s", operator)
-                .isThrownBy(() -> calculator.getResult(splitWithoutSpace(operator)))
-                .withMessage(FORMULA_NULL_ERROR.getCode() + " :: " + FORMULA_NULL_ERROR.getDesc());
-    }
-
-    static Stream<Arguments> getAlphabetsToTestWhenWrongOperatorThenThrowException() {
-        return IntStream.rangeClosed(65, 122)
-                .filter(Character::isAlphabetic)
-                .mapToObj(PostfixCalculatorTest::getAlphabetArgument);
-    }
-
-    static Arguments getAlphabetArgument(int ascii) {
-        return Arguments.arguments(valueOf((char) ascii));
-    }
-
 }
